Create flatten-array.hs (#36)
* Create flatten-array.hs

To make this problem a bit interesting in Haskell, one has to define his/her own datatype. Lists in Haskell are uniform meaning that they cannot store different datatypes (e.g., `[1, 'a', 2]` will not work). In this case, elements of sublists should also have the same datatypes meaning that `[[1,2], ['a', 'b'], [3,4]]` is also not possible. We would be left with boring lists like `[[1,2,3],[4,5,6],[7,8,9]]` where everything (on all access levels) has the same datatype so I defined the brand new - `NestedList datatype`. I also added a note regarding this datatype in `README`.

This datatype could also be found at https://wiki.haskell.org/99_questions/Solutions/7 where you can also see the similar solution.
As seen on the link above, there are lots of different approaches yet, I found this one to be one of the easiest.

* Update flatten-array.hs