Cleanup
- Changed all README files in problems to be all-upper as per convention.
- Changed the `find-missing-solutions` script:
  - Bash best practices
  - Added `-h` and `--help` support
  - Added usage text according to docopt
  - Added colored output if talking to a terminal
  - Simplified search algorithm to descend automatically

Note that the script expects a problem to have a `README.md` file in its directory. Something that is not the case for all problems. However, I believe that we need to ensure that every problem actually has a `README.md` because otherwise it is pretty useless. I added a new one to the very first Code Jam problem but we will have to add them to all our problems.