Added Test Framework
This is an attempt for standardization. We see different solutions for testing everywhere. We also see different naming everywhere. Some solutions even use dependencies (or why is there a `node_modules` ignore rule?). Most coding challenges out there are using a simple I/O approach, we should too. I think that we need to extend our contributor guides and also add some rules there:

- Single file called `solution.$ext` (where `$ext` is the official extension of the programming language).
- No dependencies except those included in the **latest** version of the language. (What about weird languages like C that have super inconsistent standard libraries across systems?)
- No error handling, input is always valid and present.

The last point might be problematic because there are problems for which it would be great to have negative tests to ensure that people actually understood the problem, e.g. #113.